pyproject.toml
[project]
name = "cognitiveos"
version = "0.1.0"
description = "Personal Cognitive Operating System - A computable external brain"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "arq>=0.26.0",
    "cashews[redis]>=7.0.0",
    "dishka>=1.8.0",
    "faiss-cpu>=1.7.0",
    "httpx>=0.27.0",
    "litellm>=1.0.0",
    "litestar>=2.21.0",
    "numpy>=1.26.0",
    "piccolo>=1.32.0",
    "piccolo-admin>=1.12.0",
    "pydantic-settings>=2.0.0",
    "redis>=5.0.0",
    "uvicorn>=0.41.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]
